Added closure constructs + list extensions
Added ActionToFuncClosure type - allows a captured Action delegate to be cast to a Func<T> that returns a given value. This allows said Action delegate to be passed to a method accepting Func<T> where the return value is not linked to the action we wish to invoke.
Added FuncArgsClosure type - allows a func with 1-8 input parameters (eg. Func<T1, TResult>) to be captured alongside its input arguments and be cast to a parameterless Func<TResult> closure.
Added ClosureExtensions class with extension methods for the types mentioned above.
Added List<T> extension methods focused around removing elements with an optional state argument.