JsonLexer supports escaped strings
Fully support all escaped chars mentioned in the json spec.
The JsonLexer now keeps a LexerState struct containing both the read index and a flag indicating whether the cursor is inside an escaped string.
Strings that contain one or more escaped chars are read via a new method ReadNextEscapedStringToken and yield tokens that represent the individual escaped char, or a span of raw string data.
Added documentation explaining this process to relevant methods, and added documentation to all TokenType enum values.
Removed redundant index incrementing when returning undefined JToken. Previously the enumerator implementation treated a zero index as a special 'unmoved' state, which caused the lexer to infinitely read the same token if an EOF or undefined token was found first (eg. json is an empty string). This issue has been resolved and as such the index does not need to advance in these cases.
Added custom test display name resolver for unit tests to improve printout of array arguments.